+++ apps/docs/src/routes/base/date-input/examples/hero.tsx
@@ -13,14 +13,12 @@ export default component$(() => {
     <div class="date-input-container">
       <DateInput.Root class="date-input-root-col">
         <DateInput.Label>My date input</DateInput.Label>
-        <DateInput.DateEntry onChange$={handleChange$}>
+        <DateInput.Entry onChange$={handleChange$} separator="/">
           <DateInput.Month />
-          <DateInput.Separator separator="/" />
           <DateInput.Day />
-          <DateInput.Separator separator="/" />
           <DateInput.Year />
           <DateInput.HiddenInput name="date" value={selectedDate.value} />
-        </DateInput.DateEntry>
+        </DateInput.Entry>
       </DateInput.Root>
 
       <div>