+++ apps/docs/src/routes/base/pagination/examples/per-page.tsx
@@ -23,9 +23,9 @@ export default component$(() => {
         const uniqueKey = `page-${index}-${Date.now()}`;
 
         return (
-          <Pagination.Page class="pagination-page" key={uniqueKey}>
+          <Pagination.Item class="pagination-item" key={uniqueKey}>
             <span>{index + 1}</span>
-          </Pagination.Page>
+          </Pagination.Item>
         );
       })}