+++ apps/docs/src/routes/base/radio-group/examples/disabled.tsx
@@ -18,24 +18,24 @@ export default component$(() => {
         <RadioGroup.Label>Choose option</RadioGroup.Label>
 
         <RadioGroup.Item value="option1" class="radio-group-item">
-          <RadioGroup.Label>Option 1</RadioGroup.Label>
-          <RadioGroup.Trigger class="radio-group-trigger">
-            <RadioGroup.Indicator class="radio-group-indicator" />
-          </RadioGroup.Trigger>
+          <RadioGroup.ItemLabel>Option 1</RadioGroup.ItemLabel>
+          <RadioGroup.ItemTrigger class="radio-group-trigger">
+            <RadioGroup.ItemIndicator class="radio-group-indicator" />
+          </RadioGroup.ItemTrigger>
         </RadioGroup.Item>
 
         <RadioGroup.Item value="option2" class="radio-group-item">
-          <RadioGroup.Label>Option 2 (Disabled)</RadioGroup.Label>
-          <RadioGroup.Trigger class="radio-group-trigger" disabled>
-            <RadioGroup.Indicator class="radio-group-indicator" />
-          </RadioGroup.Trigger>
+          <RadioGroup.ItemLabel>Option 2 (Disabled)</RadioGroup.ItemLabel>
+          <RadioGroup.ItemTrigger class="radio-group-trigger" disabled>
+            <RadioGroup.ItemIndicator class="radio-group-indicator" />
+          </RadioGroup.ItemTrigger>
         </RadioGroup.Item>
 
         <RadioGroup.Item value="option3" class="radio-group-item">
-          <RadioGroup.Label>Option 3</RadioGroup.Label>
-          <RadioGroup.Trigger class="radio-group-trigger">
-            <RadioGroup.Indicator class="radio-group-indicator" />
-          </RadioGroup.Trigger>
+          <RadioGroup.ItemLabel>Option 3</RadioGroup.ItemLabel>
+          <RadioGroup.ItemTrigger class="radio-group-trigger">
+            <RadioGroup.ItemIndicator class="radio-group-indicator" />
+          </RadioGroup.ItemTrigger>
         </RadioGroup.Item>
       </RadioGroup.Root>
     </div>