+++ libs/components/src/date-input/date-input-entry-context.tsx
@@ -1,12 +1,12 @@
-import { type Signal, createContextId } from "@builder.io/qwik";
+import { type QwikJSX, type Signal, createContextId } from "@builder.io/qwik";
 import type { ISODate } from "../calendar/types";
 import type { DateSegment } from "./types";
 
-export const dateInputDateEntryContextId = createContextId<DateInputDateEntryContext>(
-  "qds-date-input-date-entry-context"
+export const dateInputEntryContextId = createContextId<DateInputEntryContext>(
+  "qds-date-input-entry-context"
 );
 
-export type DateInputDateEntryContext = {
+export type DateInputEntryContext = {
   dateSig: Signal<ISODate | null>;
   dayOfMonthSegmentSig: Signal<DateSegment>;
   monthSegmentSig: Signal<DateSegment>;
@@ -17,4 +17,5 @@ export type DateInputDateEntryContext = {
   value?: string;
   disabledSig: Signal<boolean>;
   isInternalSegmentClearance: Signal<boolean>;
+  separator?: string | QwikJSX.Element;
 };