+++ libs/components/src/date-input/date-input-hidden-input.tsx
@@ -1,12 +1,12 @@
 import { type PropsOf, component$, useContext } from "@builder.io/qwik";
 import { VisuallyHidden } from "../visually-hidden/visually-hidden";
-import { dateInputDateEntryContextId } from "./date-input-date-entry-context";
+import { dateInputEntryContextId } from "./date-input-entry-context";
 type PublicDateInputHiddenNativeInputProps = PropsOf<"input">;
 
 /** A hidden native text input for form submission */
 export const DateInputHiddenInput = component$(
   (props: PublicDateInputHiddenNativeInputProps) => {
-    const context = useContext(dateInputDateEntryContextId);
+    const context = useContext(dateInputEntryContextId);
 
     return (
       <VisuallyHidden>