+++ libs/components/src/date-input/index.ts
@@ -1,9 +1,8 @@
 export { DateInputRoot as Root } from "./date-input-root";
-export { DateInputDateEntry as DateEntry } from "./date-input-date-entry";
+export { DateInputEntry as Entry } from "./date-input-entry";
 export { DateInputHiddenInput as HiddenInput } from "./date-input-hidden-input";
 export { DateInputLabel as Label } from "./date-input-label";
 export { DateInputSegment as Segment } from "./date-input-segment";
-export { DateInputSeparator as Separator } from "./date-input-separator";
 export { DateInputDay as Day } from "./date-input-day";
 export { DateInputMonth as Month } from "./date-input-month";
 export { DateInputYear as Year } from "./date-input-year";