+++ libs/components/src/pagination/index.ts
@@ -1,5 +1,5 @@
 export { PaginationRoot as Root } from "./pagination-root";
-export { PaginationPage as Page } from "./pagination-page";
+export { PaginationItem as Item } from "./pagination-item";
 export { PaginationNext as Next } from "./pagination-next";
 export { PaginationPrevious as Previous } from "./pagination-previous";
 export { PaginationEllipsis as Ellipsis } from "./pagination-ellipsis";