+++ libs/components/src/switch/index.ts
@@ -4,4 +4,4 @@ export { SwitchThumb as Thumb } from "./switch-thumb";
 export { SwitchLabel as Label } from "./switch-label";
 export { SwitchDescription as Description } from "./switch-description";
 export { SwitchHiddenInput as HiddenInput } from "./switch-hidden-input";
-export { SwitchErrorMessage as ErrorMessage } from "./switch-error-message";
+export { SwitchError as Error } from "./switch-error";