+++ libs/utils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @kunai-consulting/qwik-utils
 
+## 0.2.0
+
 ## 0.1.0
 
 ## 0.0.12